requirements.txt
discord_webhook==0.17.0
Flask==2.1.3
gevent_socketio==0.3.6
humanize==4.4.0
Jinja2==3.1.2
paypalrestsdk==1.13.1
python-dotenv==0.21.0
python_bcrypt==0.3.2
requests==2.28.1    